Refactor project structure, update configs, and improve cart management
