+++ mern-pro2/src/main.jsx
@@ -3,7 +3,7 @@ import { createRoot } from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App.jsx";
 import "./index.css";
-import { CartProvider } from "./context/CartContext"; // ✅ Import CartProvider
+import CartProvider from "./context/CartProvider"; // ✅ Import CartProvider
 
 // Get root element from index.html
 const rootElement = document.getElementById("root");
